+++ src/AppHeader.js
@@ -13,7 +13,8 @@ export default class AppHeader extends Component {
     return(
       <Menu className='AppHeader' color='blue' inverted stackable>
         <Menu.Item header>
-          <Image avatar src={`${process.env.PUBLIC_URL}/fleur_de_lis_black.ico`}/>
+          <Image avatar
+            src={`${process.env.PUBLIC_URL}/fleur_de_lis_black.ico`}/>
           { title }
         </Menu.Item>
         { React.Children.map(children, (child, i) =>